+++ tests/plugins/constructor_parameters/plugin.py
@@ -1,9 +1,9 @@
 from typing import Any
 
-from serpentarium import AbstractPlugin
+from serpentarium import MultiUsePlugin, NamedPluginMixin
 
 
-class Plugin(AbstractPlugin):
+class Plugin(NamedPluginMixin, MultiUsePlugin):
     def __init__(self, plugin_name: str, my_param: Any):
         super().__init__(plugin_name=plugin_name)
         self.my_param = my_param