+++ tests/plugins/plugin1/plugin.py
@@ -1,14 +1,15 @@
+import os
 import sys
 
 assert "wonderland" not in sys.modules
 
 import wonderland  # noqa: E402
 
-from serpentarium import AbstractPlugin  # noqa: E402
+from serpentarium import MultiUsePlugin, NamedPluginMixin  # noqa: E402
 
 
-class Plugin(AbstractPlugin):
+class Plugin(NamedPluginMixin, MultiUsePlugin):
     def run(self, **kwargs) -> str:
-        print("I am plugin 1")
+        print(f"I am plugin 1: {os.getpid()}")
 
         return f"My name is {wonderland.pick_twin()}"